+++ jschon/formats.py
@@ -2,8 +2,7 @@ from jschon.jsonpointer import JSONPointer
 from jschon.vocabulary.format import format_validator
 
 
-@format_validator('json-pointer')
+@format_validator('json-pointer', instance_types=('string',))
 def validate_json_pointer(value: str) -> None:
-    if isinstance(value, str):
-        if not JSONPointer._json_pointer_re.fullmatch(value):
-            raise ValueError
+    if not JSONPointer._json_pointer_re.fullmatch(value):
+        raise ValueError