+++ tests/test_formats.py
@@ -15,21 +15,28 @@ def setup_validators(catalog):
         "ipv4",
         "ipv6",
         "json-pointer",
+        "uint8",
     )
     yield
     catalog._enabled_formats.clear()
 
 
-@format_validator('ipv4')
+@format_validator('ipv4', instance_types=('string',))
 def ipv4_validator(value):
-    if isinstance(value, str):
-        ipaddress.IPv4Address(value)
+    ipaddress.IPv4Address(value)
 
 
-@format_validator('ipv6')
+@format_validator('ipv6', instance_types=('string',))
 def ipv6_validator(value):
-    if isinstance(value, str):
-        ipaddress.IPv6Address(value)
+    ipaddress.IPv6Address(value)
+
+
+@format_validator('uint8', instance_types=('number',))
+def uint8_validator(value):
+    if value % 1:
+        raise ValueError(f'{value} is not an integer (uint8)')
+    if value < 0 or value > 255:
+        raise ValueError(f'{value} is out of range for uint8')
 
 
 def evaluate(format_attr, instval, assert_=True):
@@ -86,6 +93,22 @@ def test_jsonpointer_invalid(instval):
         assert result is True
     except JSONPointerError:
         assert result is False
+
+
+@pytest.mark.parametrize('instval', (3, "1", "what"))
+def test_uint8_valid(instval):
+    result = evaluate(
+        "uint8",
+        instval,
+        isinstance(instval, (int, float))
+    )
+    assert result is True
+
+
+@pytest.mark.parametrize('instval', (3.001, -1, 256))
+def test_uint8_invalid(instval):
+    result = evaluate("uint8", instval)
+    assert result is False
 
 
 @given(instval=hs.uuids() | hs.text())